requirements.txt
pyyaml
requests==2.32.3
kubernetes==31.0.0
